requirements.txt
loguru
langchain
langchain-groq
langchain-community
langchain-openai
langchain-chroma
chromadb
sentence-transformers
pypdf
python-dotenv
pymupdf
unstructured
pdfminer
python-docx
docx2txt
pandas
openpyxl
jq
matplotlib
langgraph
rank_bm25
langchain-huggingface
langfuse
arxiv
pydantic-settings
nltk
datasets
ragas

# Database & Migrations
alembic==1.17.2
sqlmodel==0.0.22
asyncpg==0.31.0
psycopg2-binary==2.9.10

# Vector Store
pinecone-client==5.0.1

# FastAPI & Web Server (will be added in Phase 6)
fastapi
uvicorn[standard]
nest-asyncio>=1.6.0
# uvicorn[standard]==0.27.0

# Celery & Task Queue (will be added in Phase 5)
celery==5.3.6
redis==5.0.1